requirements/local.txt
cloudpickle==1.6.0
cycler==0.10.0
gym==0.18.3
gym-retro==0.8.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
opencv-python==4.5.1.48
pandas==1.2.4
Pillow==8.2.0
pyglet==1.5.15
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.3
six==1.16.0
stable-baselines3==1.0
torch==1.8.1
typing-extensions==3.10.0.0
